\name{CCES.complete}
\alias{CCES.complete}
\docType{data}
\title{
CCES.complete}
\description{
Data from CCES.
}
\usage{data(CCES.complete)}
\format{
  A data frame with 36421 observations on the following 79 variables.
  \describe{
    \item{\code{m.cca}}{a numeric vector}
    \item{\code{m.r2l}}{a numeric vector}
    \item{\code{m.naral}}{a numeric vector}
    \item{\code{m.sierra}}{a numeric vector}
    \item{\code{m.amleg}}{a numeric vector}
    \item{\code{m.vfw}}{a numeric vector}
    \item{\code{m.pta}}{a numeric vector}
    \item{\code{m.lwv}}{a numeric vector}
    \item{\code{m.nra}}{a numeric vector}
    \item{\code{m.aarp}}{a numeric vector}
    \item{\code{m.aaa}}{a numeric vector}
    \item{\code{fiscal0.borrow}}{a numeric vector}
    \item{\code{fiscal0.tax}}{a numeric vector}
    \item{\code{fiscal0.domestic}}{a numeric vector}
    \item{\code{fiscal0.defense}}{a numeric vector}
    \item{\code{fiscal1.borrow}}{a numeric vector}
    \item{\code{fiscal1.tax}}{a numeric vector}
    \item{\code{fiscal1.domestic}}{a numeric vector}
    \item{\code{fiscal1.defense}}{a numeric vector}
    \item{\code{stax.income}}{a numeric vector}
    \item{\code{sbudget.morecuts}}{a numeric vector}
    \item{\code{pid2}}{a numeric vector}
    \item{\code{selfplace.2}}{a numeric vector}
    \item{\code{ssen.rvote}}{a numeric vector}
    \item{\code{srep.rvote}}{a numeric vector}
    \item{\code{streas.rvote}}{a numeric vector}
    \item{\code{sstate.rvote}}{a numeric vector}
    \item{\code{atgen.rvote}}{a numeric vector}
    \item{\code{ltgov.rvote}}{a numeric vector}
    \item{\code{hou.rvote}}{a numeric vector}
    \item{\code{sen.rvote}}{a numeric vector}
    \item{\code{governor.rvote}}{a numeric vector}
    \item{\code{cafta}}{a numeric vector}
    \item{\code{capgain.cut}}{a numeric vector}
    \item{\code{minwage}}{a numeric vector}
    \item{\code{imm.citizen.path}}{a numeric vector}
    \item{\code{iraq.timetable.yes}}{a numeric vector}
    \item{\code{fund.stemcells}}{a numeric vector}
    \item{\code{ban.late.abort}}{a numeric vector}
    \item{\code{selfplace.1}}{a numeric vector}
    \item{\code{mil.none}}{a numeric vector}
    \item{\code{mil.help.un}}{a numeric vector}
    \item{\code{mil.protect.allies}}{a numeric vector}
    \item{\code{mil.spread.dem}}{a numeric vector}
    \item{\code{mil.genocide}}{a numeric vector}
    \item{\code{mil.terrcamp}}{a numeric vector}
    \item{\code{mil.oil}}{a numeric vector}
    \item{\code{aff.act}}{a numeric vector}
    \item{\code{ss.priv}}{a numeric vector}
    \item{\code{env.over.jobs}}{a numeric vector}
    \item{\code{pro.choice}}{a numeric vector}
    \item{\code{rep.approve}}{a numeric vector}
    \item{\code{sen2.approve}}{a numeric vector}
    \item{\code{sen1.approve}}{a numeric vector}
    \item{\code{governor.approve}}{a numeric vector}
    \item{\code{iraq.mistake}}{a numeric vector}
    \item{\code{state.econ.pastyear.better}}{a numeric vector}
    \item{\code{nat.econ.pastyear.better}}{a numeric vector}
    \item{\code{pid7.r}}{a numeric vector}
    \item{\code{pid3.r}}{a numeric vector}
    \item{\code{local.econ.good}}{a numeric vector}
    \item{\code{fox.biased}}{a numeric vector}
    \item{\code{ban.gaymarr}}{a numeric vector}
    \item{\code{stemcells.favor}}{a numeric vector}
    \item{\code{imm.enforce}}{a numeric vector}
    \item{\code{iraq.neverstable}}{a numeric vector}
    \item{\code{iraq.mistake.1}}{a numeric vector}
    \item{\code{climatescience.deny}}{a numeric vector}
    \item{\code{rcontrol}}{a numeric vector}
    \item{\code{generic.rvote}}{a numeric vector}
    \item{\code{minwage.1}}{a numeric vector}
    \item{\code{sex}}{a factor with levels \code{Male} \code{Female}}
    \item{\code{age}}{an ordered factor with levels \code{18-29} < \code{30-44} < \code{45-64} < \code{65+}}
    \item{\code{z.age}}{a numeric vector}
    \item{\code{income}}{an ordered factor with levels \code{less than $20,000} < \code{$20,000-$40,000} < \code{$40,000-$75,000} < \code{$75,000-$150,000} < \code{$150,000+}}
    \item{\code{z.income}}{a numeric vector}
    \item{\code{education}}{an ordered factor with levels \code{less than a high school education} < \code{high school graduate} < \code{some college} < \code{college graduate} < \code{postgraduate}}
    \item{\code{race}}{a factor with levels \code{White} \code{Black} \code{Hispanic} \code{Other}}
    \item{\code{state}}{a factor with levels \code{AK} \code{AL} \code{AR} \code{AZ} \code{CA} \code{CO} \code{CT} \code{DC} \code{DE} \code{FL} \code{GA} \code{HI} \code{IA} \code{ID} \code{IL} \code{IN} \code{KS} \code{KY} \code{LA} \code{MA} \code{MD} \code{ME} \code{MI} \code{MN} \code{MO} \code{MS} \code{MT} \code{NC} \code{ND} \code{NE} \code{NH} \code{NJ} \code{NM} \code{NV} \code{NY} \code{OH} \code{OK} \code{OR} \code{PA} \code{RI} \code{SC} \code{SD} \code{TN} \code{TX} \code{UT} \code{VA} \code{VT} \code{WA} \code{WI} \code{WV} \code{WY}}
  }
}

\examples{
data(CCES.complete)
## maybe str(CCES.complete)
}
\keyword{datasets}
