\name{mrp.regions}
\alias{mrp.regions}
\docType{data}
\title{
States and Regions for MRP
}
\description{
A data.frame containing the regions for the 50 states plus the District of Columbia, indexed by state abbreviation.
}
\usage{data(mrp.regions)}
\format{
  A data frame with 51 observations on the following 2 variables.
  \describe{
    \item{\code{state}}{a factor with levels \code{AK} \code{AL} \code{AR} \code{AZ} \code{CA} \code{CO} \code{CT} \code{DC} \code{DE} \code{FL} \code{GA} \code{HI} \code{IA} \code{ID} \code{IL} \code{IN} \code{KS} \code{KY} \code{LA} \code{MA} \code{MD} \code{ME} \code{MI} \code{MN} \code{MO} \code{MS} \code{MT} \code{NC} \code{ND} \code{NE} \code{NH} \code{NJ} \code{NM} \code{NV} \code{NY} \code{OH} \code{OK} \code{OR} \code{PA} \code{RI} \code{SC} \code{SD} \code{TN} \code{TX} \code{UT} \code{VA} \code{VT} \code{WA} \code{WI} \code{WV} \code{WY}}
    \item{\code{region}}{a factor with levels \code{DC} \code{Midwest} \code{Northeast} \code{South} \code{West}}
  }
}

\examples{
data(mrp.regions)
}
\keyword{datasets}
