\name{samesexmarriage}
\alias{marriage.data}
\alias{Statelevel}
\alias{samesexmarriage}
\docType{data}
\title{
Example for MRP -- combined 2004-2005 poll data: support for same-sex civil marriage}
\description{
Example dataset for MRP: poll data on support for same-sex civil
marriage in the United States, with state-level predictors. See the
example in \link{mrp}.
}
\usage{data(samesexmarriage)}
\format{
  \subsection{marriage.data}{
    A data frame with 6548 observations on the following 27 variables.
  \describe{
    \item{\code{poll}}{a character vector}
    \item{\code{poll.firm}}{a character vector}
    \item{\code{poll.year}}{a numeric vector}
    \item{\code{id}}{a numeric vector}
    \item{\code{statenum}}{a numeric vector}
    \item{\code{statename}}{a factor with levels \code{alabama} \code{alaska} \code{arizona} \code{arkansas} \code{california} \code{colorado} \code{connecticut} \code{delaware} \code{district of columbia} \code{florida} \code{georgia} \code{hawaii} \code{idaho} \code{illinois} \code{indiana} \code{iowa} \code{kansas} \code{kentucky} \code{louisiana} \code{maine} \code{maryland} \code{massachusetts} \code{michigan} \code{minnesota} \code{mississippi} \code{missouri} \code{montana} \code{nebraska} \code{nevada} \code{new hampshire} \code{new jersey} \code{new mexico} \code{new york} \code{north carolina} \code{north dakota} \code{ohio} \code{oklahoma} \code{oregon} \code{pennsylvania} \code{rhode island} \code{south carolina} \code{south dakota} \code{tennessee} \code{texas} \code{utah} \code{vermont} \code{virginia} \code{washington} \code{west virginia} \code{wisconsin} \code{wyoming}}
    \item{\code{region.cat}}{a numeric vector}
    \item{\code{female}}{a numeric vector}
    \item{\code{race.wbh}}{a numeric vector}
    \item{\code{edu.cat}}{a numeric vector}
    \item{\code{age.cat}}{a numeric vector}
    \item{\code{age.cat6}}{a numeric vector}
    \item{\code{age.edu.cat6}}{a numeric vector}
    \item{\code{educ}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{democrat}}{a numeric vector}
    \item{\code{republican}}{a numeric vector}
    \item{\code{black}}{a numeric vector}
    \item{\code{hispanic}}{a numeric vector}
    \item{\code{weight}}{a numeric vector}
    \item{\code{yes.of.opinion.holders}}{a numeric vector}
    \item{\code{yes.of.all}}{a numeric vector}
    \item{\code{state}}{a character vector}
    \item{\code{state.initnum}}{a numeric vector}
    \item{\code{region}}{a character vector}
    \item{\code{no.of.all}}{a numeric vector}
    \item{\code{no.of.opinion.holders}}{a numeric vector}
  }}
  \subsection{Statelevel}{
    A data frame with 51 observations on the following 50 variables.
  \describe{
    \item{\code{sstate.initnum}}{a numeric vector}
    \item{\code{sstate}}{a character vector}
    \item{\code{sstatename}}{a character vector}
    \item{\code{ideology}}{a numeric vector}
    \item{\code{professional.index}}{a numeric vector}
    \item{\code{professional.cat}}{a numeric vector}
    \item{\code{salary}}{a numeric vector}
    \item{\code{totalday}}{a numeric vector}
    \item{\code{staffper}}{a numeric vector}
    \item{\code{citizen.init}}{a numeric vector}
    \item{\code{year.adoption}}{a numeric vector}
    \item{\code{consitutional.init}}{a numeric vector}
    \item{\code{signature}}{a numeric vector}
    \item{\code{usage}}{a numeric vector}
    \item{\code{legislative.insulation}}{a numeric vector}
    \item{\code{p.evang}}{a numeric vector}
    \item{\code{p.mormon}}{a numeric vector}
    \item{\code{p.cath}}{a numeric vector}
    \item{\code{p.metroarea}}{a numeric vector}
    \item{\code{adjinc.cap}}{a numeric vector}
    \item{\code{pblack}}{a numeric vector}
    \item{\code{phisp}}{a numeric vector}
    \item{\code{p65over}}{a numeric vector}
    \item{\code{phighsch}}{a numeric vector}
    \item{\code{pbachelors}}{a numeric vector}
    \item{\code{demcontrol}}{a numeric vector}
    \item{\code{pdemocrat}}{a numeric vector}
    \item{\code{polarization}}{a numeric vector}
    \item{\code{divided}}{a numeric vector}
    \item{\code{kerry.04}}{a numeric vector}
    \item{\code{gore.00}}{a numeric vector}
    \item{\code{clinton.92}}{a numeric vector}
    \item{\code{mondale.84}}{a numeric vector}
    \item{\code{dukakis.88}}{a numeric vector}
    \item{\code{humphrey.68}}{a numeric vector}
    \item{\code{mcgovern.72}}{a numeric vector}
    \item{\code{population}}{a numeric vector}
    \item{\code{court.type}}{a character vector}
    \item{\code{court.elected}}{a numeric vector}
    \item{\code{berry.citizen}}{a numeric vector}
    \item{\code{berry.government}}{a numeric vector}
    \item{\code{Bryan.1896}}{a numeric vector}
    \item{\code{tempculture}}{a character vector}
    \item{\code{state.culture}}{a character vector}
    \item{\code{p.kerry.full}}{a numeric vector}
    \item{\code{p.relig.full}}{a numeric vector}
    \item{\code{p.mormon.full}}{a numeric vector}
    \item{\code{p.evang.full}}{a numeric vector}
    \item{\code{state.initnum}}{a numeric vector}
    \item{\code{state}}{a character vector}
}}
}
%\details{
%}
\source{
Adapted from \url{http://www.princeton.edu/~jkastell/mrp_primer.html}}
\references{
See also \url{http://www.nytimes.com/2010/08/22/weekinreview/22gay.html?_r=1}}
%\examples{
%}
\keyword{datasets}


