\name{Statelevel}
\alias{Statelevel}
\docType{data}
\title{
Example for MRP: Assorted state-level predictors}
\description{
Example dataset for MRP: state-level predictors. See the
example in \link{mrp}.
}
\usage{data(samesexmarriage)}
\format{
  \subsection{Statelevel}{
    A data frame with 51 observations on the following 50 variables.
  \describe{
    \item{\code{sstate.initnum}}{a numeric vector}
    \item{\code{sstate}}{a character vector}
    \item{\code{sstatename}}{a character vector}
    \item{\code{ideology}}{a numeric vector}
    \item{\code{professional.index}}{a numeric vector}
    \item{\code{professional.cat}}{a numeric vector}
    \item{\code{salary}}{a numeric vector}
    \item{\code{totalday}}{a numeric vector}
    \item{\code{staffper}}{a numeric vector}
    \item{\code{citizen.init}}{a numeric vector}
    \item{\code{year.adoption}}{a numeric vector}
    \item{\code{consitutional.init}}{a numeric vector}
    \item{\code{signature}}{a numeric vector}
    \item{\code{usage}}{a numeric vector}
    \item{\code{legislative.insulation}}{a numeric vector}
    \item{\code{p.evang}}{a numeric vector}
    \item{\code{p.mormon}}{a numeric vector}
    \item{\code{p.cath}}{a numeric vector}
    \item{\code{p.metroarea}}{a numeric vector}
    \item{\code{adjinc.cap}}{a numeric vector}
    \item{\code{pblack}}{a numeric vector}
    \item{\code{phisp}}{a numeric vector}
    \item{\code{p65over}}{a numeric vector}
    \item{\code{phighsch}}{a numeric vector}
    \item{\code{pbachelors}}{a numeric vector}
    \item{\code{demcontrol}}{a numeric vector}
    \item{\code{pdemocrat}}{a numeric vector}
    \item{\code{polarization}}{a numeric vector}
    \item{\code{divided}}{a numeric vector}
    \item{\code{kerry.04}}{a numeric vector}
    \item{\code{gore.00}}{a numeric vector}
    \item{\code{clinton.92}}{a numeric vector}
    \item{\code{mondale.84}}{a numeric vector}
    \item{\code{dukakis.88}}{a numeric vector}
    \item{\code{humphrey.68}}{a numeric vector}
    \item{\code{mcgovern.72}}{a numeric vector}
    \item{\code{population}}{a numeric vector}
    \item{\code{court.type}}{a character vector}
    \item{\code{court.elected}}{a numeric vector}
    \item{\code{berry.citizen}}{a numeric vector}
    \item{\code{berry.government}}{a numeric vector}
    \item{\code{Bryan.1896}}{a numeric vector}
    \item{\code{tempculture}}{a character vector}
    \item{\code{state.culture}}{a character vector}
    \item{\code{p.kerry.full}}{a numeric vector}
    \item{\code{p.relig.full}}{a numeric vector}
    \item{\code{p.mormon.full}}{a numeric vector}
    \item{\code{p.evang.full}}{a numeric vector}
    \item{\code{state.initnum}}{a numeric vector}
    \item{\code{state}}{a character vector}
}}
}
%\details{
%}
\source{
Adapted from \url{http://www.princeton.edu/~jkastell/mrp_primer.html}}
\references{
See also \url{http://www.nytimes.com/2010/08/22/weekinreview/22gay.html?_r=1}}
%\examples{
%}
\keyword{datasets}


